viz: pull audio stuff into a node
This is still not a graph of any kind.

Observe, the App is extremely lightweight.  This is what we wanted.  How we
build nodes and connect them is becoming decoupled from the frontend.

Most nodes should find their way into lib in preparation for multi-frontend
support.

Development of a render graph style data structure will enable mixing and
matching nodes.  This is also preparation for scripting "presets".

Machine learning will tee some of these nodes in order to enable online learning.